change constant max Execution Proof subnets to long to avoid unnecessary conversions

+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/gossip/topics/GossipTopics.java
@@ -118,7 +118,7 @@ public class GossipTopics {
     topics.addAll(getAllDataColumnSidecarSubnetTopics(gossipEncoding, forkDigest, spec));
 
     if (p2pConfig.isExecutionProofTopicEnabled()) {
-      for (int i = 0; i < Constants.MAX_EXECUTION_PROOF_SUBNETS.intValue(); i++) {
+      for (int i = 0; i < Constants.MAX_EXECUTION_PROOF_SUBNETS; i++) {
         topics.add(getExecutionProofSubnetTopic(forkDigest, i, gossipEncoding));
       }
     }
